#1 - Refactor classes
