+++ src/models.py
@@ -1,0 +1,6 @@
+# src/api/models.py
+
+class Users:
+    def __init__(self, id, balance) -> None:
+        self.id = id
+        self.balance = balance
